requirements.txt
FastAPI
pydantic
typing
openai
SQLAlchemy
pymysql
uvicorn
cloudinary
pymongo
email-validator
python-jose[cryptography]
passlib[bcrypt]
motor
FastAPI
pydantic
typing
openai
SQLAlchemy
pymysql
uvicorn
cloudinary